requirements.txt
streamlit
PyPDF2
langchain-google-genai==2.0.10
python-dotenv
langchain==0.3.19
langchain-community==0.3.18